Initial HUD implementation
Added player and enemy health bars.
Changed GUI scene to no longer be auto-loaded. Therefore, it needs to be manually added to any level/scene that needs to display GUI. This was required because the GUI needs game state, which is not available in the main menu.
Added health component to player. Enemies can now damage player.
Enabled anti-aliasing.